Добавлен диструктор у родительского класса

+++ khismatullin.kamil/A1/main.cpp
@@ -3,7 +3,7 @@
 #include "circle.hpp"
 
 int main() {
-  Rectangle *rectangle = new Rectangle{{(double) 2, (double) 4}, 7, 10};  //создание прямоугольника
+  Rectangle *rectangle = new Rectangle{{2, 4}, 7, 10};  //создание прямоугольника
   std::cout << "[Rectangle]" << std::endl;  //вывод параметров прямоугольника(на следующей строке)
   std::cout << "Center : (" << rectangle->getCenter().x << " , " << rectangle->getCenter().y << ")" << std::endl
       << "Height : " << rectangle->getFrameRect().height << std::endl
@@ -11,15 +11,15 @@ int main() {
       << "Area   : " << rectangle->getArea() << std::endl;
 
   std::cout << std::endl << "{Moving}" << std::endl;
-  rectangle->move({(double) -3, (double) -3});  //перемещение по координатам и дальнейший вывод текущего положения
+  rectangle->move({-3, -3});  //перемещение по координатам и дальнейший вывод текущего положения
   std::cout << "-- to point : (" << rectangle->getCenter().x << " , " << rectangle->getCenter().y << ")" << std::endl;
-  rectangle->move((double) 3, (double) 3);  //перемещение по смещению и дальнейший вывод текущего положения
+  rectangle->move(3, 3);  //перемещение по смещению и дальнейший вывод текущего положения
   std::cout << "-- by offset : (" << rectangle->getCenter().x << " , " << rectangle->getCenter().y << ")" << std::endl;
   delete (rectangle);
 
   std::cout << "___________________________" << std::endl;
 
-  Circle *circle = new Circle{{(double) 2, (double) 4}, 5}; //создание круга
+  Circle *circle = new Circle{{2, 4}, 5}; //создание круга
   std::cout << "[Circle]" << std::endl; //вывод параметров круга(на следующей строке)
   std::cout << "Center : (" << circle->getCenter().x << " , " << circle->getCenter().y << ")" << std::endl
       << "Radius : " << circle->getRadius() << std::endl
@@ -28,9 +28,9 @@ int main() {
       << "Area   : " << circle->getArea() << std::endl;
 
   std::cout << std::endl << "{Moving}" << std::endl;
-  circle->move({(double) 3, (double) 3}); //перемещение по координатам и дальнейший вывод текущего положения
+  circle->move({3, 3}); //перемещение по координатам и дальнейший вывод текущего положения
   std::cout << "-- to point : (" << circle->getCenter().x << " , " << circle->getCenter().y << ")" << std::endl;
-  circle->move((double) 3, (double) 3); //перемещение по смещению и дальнейший вывод текущего положения
+  circle->move(3, 3); //перемещение по смещению и дальнейший вывод текущего положения
   std::cout << "-- by offset : (" << circle->getCenter().x << " , " << circle->getCenter().y << ")" << std::endl;
   delete (circle);
